+++ consensus/common_test.go
@@ -242,7 +242,7 @@ func newConsensusStateWithConfig(thisConfig *cfg.Config, state *sm.State, pv *ty
 	// Make Mempool
 	mempool := mempl.NewMempool(thisConfig.Mempool, proxyAppConnMem, 0)
 	mempool.SetLogger(log.TestingLogger().With("module", "mempool"))
-	if thisConfig.Consensus.NoEmptyBlocks {
+	if thisConfig.Consensus.WaitForTxs() {
 		mempool.EnableTxsAvailable()
 	}